pyproject.toml
[project]
name = "ai-multimodal-storyteller"
version = "0.1.0"
description = "Multimodal storytelling pipeline with LLM, Stable Diffusion, TTS, MusicGen, and LangChain."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core AI/ML
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "torchaudio>=2.5.1",
    "transformers>=4.56.1",
    "diffusers>=0.35.1",
    "accelerate>=1.10.1",
    "safetensors>=0.6.2",
    
    # Audio Processing
    "coqui-tts>=0.27.1",
    "moviepy>=2.2.1",
    
    # Text Processing & Embeddings
    "sentence-transformers>=5.1.0",
    "langchain>=0.3.27",
    "chromadb>=1.1.0",
    
    # Web Interface
    "streamlit>=1.49.1",
    
    # Utilities
    "python-dotenv>=1.1.1",
    "numpy>=2.1.2",
    "pillow>=11.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"